functional-tester/tester: clean up stress, fix stress closing
